requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
distro==1.9.0
et-xmlfile==1.1.0
JPype1==1.5.0
Levenshtein==0.25.1
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20240706
pillow==10.4.0
poppler-utils==0.1.0
pycparser==2.22
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
rapidfuzz==3.9.5
six==1.16.0
tabula-py==2.9.3
tqdm==4.66.4
tzdata==2024.1